refactor: 카트 관련 순수 함수들을 models/cart.ts로 분리
- 관련함수 cart.ts 모듈로 이동
- App.tsx에서 카트 관련 함수들을 import로 변경
- 사용하지 않는 Product import 제거
- TODO 주석 정리 및 구조화
